Shows all the information from the backend
