+++ src/components/Card/Card.scss
@@ -243,7 +243,7 @@
     @include desktop {
       height: 182px;
       width: 232px;
-      padding: 32px 37px 40px 20px;
+      padding: 0 37px 40px 20px;
     }
 
     .card-title {
