+++ src/redux-toolkit/adapters/dishAdapter.ts
@@ -1,4 +1,7 @@
 import { Section } from "../../models/types";
+import spicyFoodIcon from "../../assets/images/spicyFoodIcon.svg";
+import veganFoodIcon from "../../assets/images/veganFoodIcon.svg";
+import vegiFoodIcon from "../../assets/images/vegiFoodIcon.svg";
 
 interface BackendDish {
   _id: string;
@@ -10,12 +13,25 @@ interface BackendDish {
   restaurant: string;
 }
 
+const getFoodIcon = (foodIcon: string): string | undefined => {
+  switch (foodIcon) {
+    case "Spicy":
+      return spicyFoodIcon;
+    case "Vegan":
+      return veganFoodIcon;
+    case "Vegi":
+      return vegiFoodIcon;
+    default:
+      return undefined;
+  }
+};
+
 export const transformDishData = (data: BackendDish[]): Section => {
   const cards = data.map((dish) => ({
     title: dish.title,
     image: dish.image,
     description: dish.ingredients.join(", "),
-    foodIcon: dish.tags[0],
+    foodIcon: getFoodIcon(dish.tags[0]),
     price: dish.price,
   }));
